requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
python-dotenv==1.0.0
pymongo==4.5.0
motor==3.3.1
firebase-admin==6.2.0
python-multipart==0.0.6
cloudinary==1.34.0
transformers==4.33.1
torch==2.0.1
pillow==10.0.1
huggingface-hub==0.16.4
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
httpx==0.24.1
